Add trustcacert option to Keystore for pskey
If you are adding a Certificate Authority root to pskey you want it to be
trusted

per module docs:
Certificate authorities input into a keystore aren’t trusted by default,
so if you are adding a CA you need to set this parameter to 'true'.
Valid options: 'true' or 'false'. Default: 'false'.